src/test/java/ismagicsquare/MagicTest.java
package ismagicsquare;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class MagicTest {

    @Test
    void isMagicSquare() {
        Magic m = new Magic();

        int[][] a = {{2, 7, 6}, {9, 5, 1}, {4, 3, 8}};
        int[][] b = {{3, 3, 3}, {3, 3, 3}, {3, 3, 3}};
        int[][] c = {{3, 3, 3}, {1, 2, 3}, {3, 2, 1}};
        int[][] d = {{3, 3}, {1, 2, 3}, {3, 2, 1}};

        assertEquals(true, m.isMagicSquare(a));
        assertEquals(true, m.isMagicSquare(b));
        assertEquals(false, m.isMagicSquare(c));
        assertEquals(false, m.isMagicSquare(d));
    }
}